+++ scripts/lichessTools-init.js
@@ -69,6 +69,9 @@ lichessTools.loadTool(LiChessTools.Tools.StudyLearnFromMistakes);
 lichessTools.loadTool(LiChessTools.Tools.PinOnMainPage);
 lichessTools.loadTool(LiChessTools.Tools.PgnEditor);
 lichessTools.loadTool(LiChessTools.Tools.PlayerLag);
+lichessTools.loadTool(LiChessTools.Tools.ProfileSliderText);
+lichessTools.loadTool(LiChessTools.Tools.FixCoords);
+lichessTools.loadTool(LiChessTools.Tools.DownloadAllStudies);
 //lichessTools.loadTool(LiChessTools.Tools.DownloadGames);
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);